moved localdevauth api calls into authenticated and unauthenticated files to avoid circular deps
