+++ packages/cms-lib/api/localDevAuth/authenticated.js
@@ -1,4 +1,4 @@
-const http = require('../http');
+const http = require('../../http');
 
 async function fetchScopeData(portalId, scopeGroup) {
   return http.get(portalId, {
